API/requirements.txt
mysql-connector-python
requests